# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for Schej.it!",
        "title": "Schej.it API",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "host": "localhost:3002/api",
    "paths": {
...
        "/analytics/user/{email}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Gets the user by email",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User email",
                        "name": "email",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                }
            }
        },
...
        "/users": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Returns users that match the search query",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query matching users' names/emails",
                        "name": "query",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An array of user profile objects",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.User"
                            }
                        }
                    }
                }
            }
        },
        "/users/{userId}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Returns the user by their user id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A user profile object",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.AppleCalendarAuth": {
            "type": "object"
        },
...
        "models.BufferTimeOptions": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                },
                "time": {
                    "type": "integer"
                }
            }
        },
        "models.CalendarAccount": {
            "type": "object",
            "properties": {
                "appleCalendarAuth": {
                    "type": "object",
                    "$ref": "#/definitions/models.AppleCalendarAuth"
                },
                "calendarType": {
                    "type": "string"
                },
                "email": {
                    "description": "Email is required for all calendar accounts",
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "oAuth2CalendarAuth": {
                    "type": "object",
                    "$ref": "#/definitions/models.OAuth2CalendarAuth"
                },
                "picture": {
                    "type": "string"
                },
                "subCalendars": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/models.SubCalendar"
                    }
                }
            }
        },
...
        "models.CalendarOptions": {
            "type": "object",
            "properties": {
                "bufferTime": {
                    "type": "object",
                    "$ref": "#/definitions/models.BufferTimeOptions"
                },
                "workingHours": {
                    "type": "object",
                    "$ref": "#/definitions/models.WorkingHoursOptions"
                }
            }
        },
...
        "models.OAuth2CalendarAuth": {
            "type": "object"
        },
...
        "models.SubCalendar": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "_id": {
                    "description": "Profile info",
                    "type": "string"
                },
                "calendarAccounts": {
                    "description": "CalendarAccounts is a mapping from {`email_CALENDARTYPE` =\u003e CalendarAccount} that contains all the\nadditional accounts the user wants to see google calendar events for",
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/models.CalendarAccount"
                    }
                },
                "calendarOptions": {
                    "description": "Calendar options",
                    "type": "object",
                    "$ref": "#/definitions/models.CalendarOptions"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "hasCustomName": {
                    "description": "Whether the user has set a custom name for themselves, i.e. don't change their name when they sign in",
                    "type": "boolean"
                },
                "isPremium": {
                    "type": "boolean"
                },
                "lastName": {
                    "type": "string"
                },
                "numEventsCreated": {
                    "type": "integer"
                },
                "picture": {
                    "type": "string"
                },
                "primaryAccountKey": {
                    "description": "The calendarAccountKey of the account the user first signed in with",
                    "type": "string"
                },
                "timezoneOffset": {
                    "type": "integer"
                }
            }
        },
        "models.WorkingHoursOptions": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                },
                "endTime": {
                    "type": "number"
                },
                "startTime": {
                    "type": "number"
                }
            }
        },
...
    }
}